requirements.txt
absl-py==2.1.0
altgraph==0.17.4
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
efficientnet==1.1.1
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
google-auth==2.31.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.7
imageio==2.34.2
jax==0.4.30
jaxlib==0.4.30
keras==2.12.0
Keras-Applications==1.0.8
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
macholib==1.16.3
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.9.1
ml-dtypes==0.4.0
mysql-connector-python==9.0.0
networkx==3.3
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.10.0.84
opt-einsum==3.3.0
packaging==24.1
pillow==10.4.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pyinstaller==6.9.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.23.2
scipy==1.14.0
six==1.16.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
tifffile==2024.7.2
tk==0.1.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.14.1
mysql-connector-python-rf==2.2.2
matplotlib==3.7.2
matplotlib-inline==0.1.6
threadpoolctl==3.4.0